LayoutTests/inspector-protocol/media-query-listener-exception-expected.txt

CONSOLE MESSAGE: line 13: 
CONSOLE MESSAGE: line 13: 
Test that uncaught exception in MediaQueryListListener will be reported to the console. On success you should see two exceptions in the listener logged to the console (first time when the media type is overridden and second - when they are restored). Bug 105162.
